2. Main program body The body for a program named "Example" is as follows: public class Example {     public static void main(String[] args) {         // program code     } }
The program is stored in a text file named after the program with the .java extension. For a program named Example, the file should be named Example.java.
The execution of the program begins at the part marked with the // program code comment above. During our first week of programming, we will limit ourselves to this part. When we are talking about commands such as printing, we need to write the commands into the program body. For example: System.out.print("Text to be printed");