Exercise 42.1: Loops and ending loops Create a loop in the program that works as a base for the rest of the user interface. Ask the user to submit the command inside the loop. If the command is "quit", break the loop. Use the command hangman.gameOn as the condition for the while structure. The loop should look like: while (hangman.gameOn()) {     String command = reader.nextLine();     // ... }
In the next set (week) of exercises, we will find out what this peculiar-looking condition for ending the loop is about.
This far the program should produce the following output:
************
* Hangman *
************

 * menu *
quit   - quits the game
status - prints the game status
a single letter uses the letter as a guess
an empty line prints this menu

Type a command:
do not quit

Type a command:
quit
Thank you for playing!