Exercise 42.3: Making a guess If the user types in a single letter as a command, use it to make a guess. Guessing a letter occurs in the method hangman.guess(command). The guessing command has its own printing functionality, which it uses to print more information about the guess. Hint: finding out if the command is a single letter is done as follows: String command = reader.nextLine();
if(command.length() == 1) {  // command has only one letter, so it must be a guess
    hangman.guess(command);
}

...
Type a command:
a
The letter a is not in the word.

Type a command:
b
The letter b is not in the word.

Type a command:
c
The letter c was found in the word!

Type a command:
quit
Thank you for playing!